[r4123] Изменение : MpaDecFilter - при бистриме DTS-HD HRA создаем выходной MediaType как 2 канальный. Это исправит проблему с воспроизведением на некоторых A/V ресиверах.

+++ src/filters/transform/MpaDecFilter/MpaDecFilter.h
@@ -125,7 +125,7 @@ protected:
 	BOOL            m_bFlushing;
 	BOOL            m_bNeedSyncPoint;
 
-	int             m_DTSHDProfile;
+	BYTE            m_DTSHDProfile;
 
 	REFERENCE_TIME  m_rtStartInput;
 	REFERENCE_TIME  m_rtStopInput;
